requirements.txt
python-dotenv
notion-client
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
streamlit
tiktoken
boto3